requirements.txt
fastapi
uvicorn
python-multipart
pandas
aiofiles
python-dotenv
requests
